Backend/ml/anomaly/requirements.txt
pymongo
python-dotenv
scikit-learn
pandas
numpy
requests
twilio
